todoapp_frontend/pyproject.toml
[tool.poetry]
name = "todoapp-frontend"
version = "0.1.0"
description = ""
authors = ["APAC\\HTR5HC <truc.hoangnguyen@vn.bosch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
requests-mock = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
